GH-26 further refactor
