+++ bot.Dockerfile
