+++ captain-definition-bot
@@ -1,4 +1,4 @@
 {
     "schemaVersion": 2,
-    "dockerfilePath": "./Dockerfile.bot"
+    "dockerfilePath": "./bot.Dockerfile"
 }
