+++ captain-definition-rest-api
@@ -1,4 +1,4 @@
 {
     "schemaVersion": 2,
-    "dockerfilePath": "./Dockerfile.rest-api"
+    "dockerfilePath": "./rest-api.Dockerfile"
 }
