--- ctfdb/src/ctfd/mod.rs
@@ -1,3 +1,0 @@
-pub mod api;
-pub mod structs;
-pub mod db;
