+++ ctfdb/src/ctfs/ctfd/mod.rs
@@ -1,0 +1,1 @@
+pub mod api;
