+++ ctfdb/src/ctfs/mod.rs
@@ -1,0 +1,3 @@
+pub mod db;
+pub mod ctfd;
+pub mod structs;
