+++ ctfdb/src/ctfs/structs.rs
