+++ rest-api/src/main.rs
@@ -4,7 +4,7 @@ use actix_web::{get, middleware, web, App, HttpResponse, HttpServer, Responder};
 
 use actix_cors::Cors;
 use chrono::Utc;
-use ctfdb::ctfd::db::{get_active_ctfs, get_challenges_for_ctfid, get_latest_scoreboard_status};
+use ctfdb::ctfs::db::{get_active_ctfs, get_challenges_for_ctfid, get_latest_scoreboard_status};
 use std::env;
 
 use serde::Serialize;
